moved classes like Stats, Account and Options a new file called definitions
