+++ harvest/api/kraken.py
@@ -9,6 +9,7 @@ import pandas as pd
 
 # Submodule imports
 from harvest.api._base import API
+from harvest.definitions import *
 from harvest.utils import *
