+++ harvest/api/paper.py
@@ -13,6 +13,7 @@ import yaml
 # Submodule imports
 from harvest.api._base import API
 from harvest.api.dummy import DummyStreamer
+from harvest.definitions import *
 from harvest.utils import *
