+++ harvest/api/polygon.py
@@ -10,6 +10,7 @@ from polygon import RESTClient
 
 # Submodule imports
 from harvest.api._base import API
+from harvest.definitions import *
 from harvest.utils import *
