+++ tests/livetest/test_broker_common.py
@@ -7,6 +7,7 @@ from harvest.api.kraken import Kraken
 # from harvest.api.paper import PaperBroker
 from harvest.api.yahoo import YahooStreamer
 
+from harvest.definitions import *
 from harvest.utils import *
 import time
 import os
