+++ tests/unittest/test_api_dummy.py
@@ -7,6 +7,7 @@ import pytz
 from pandas.testing import assert_frame_equal
 
 from harvest.api.dummy import DummyStreamer
+from harvest.definitions import *
 from harvest.utils import *
