+++ tests/unittest/test_api_yahoo.py
@@ -5,6 +5,7 @@ import datetime as dt
 
 from harvest.api.yahoo import YahooStreamer
 from harvest.trader.trader import PaperTrader
+from harvest.definitions import *
 from harvest.utils import *
