+++ tests/unittest/test_storage_csv.py
@@ -7,6 +7,7 @@ import pandas as pd
 from pandas.testing import assert_frame_equal
 
 from harvest.storage import CSVStorage
+from harvest.definitions import *
 from harvest.utils import *
