+++ tests/unittest/test_tester.py
@@ -7,6 +7,7 @@ import shutil
 from harvest.trader import BackTester
 from harvest.algo import BaseAlgo
 from harvest.api.dummy import DummyStreamer
+from harvest.definitions import *
 from harvest.utils import *
